renamed AppProvider.js to AppContext.js and moved on context folder
