+++ src/components/SnippetMenu.js
@@ -1,5 +1,5 @@
 import React, { useState, useContext } from "react";
-import { AppContext } from "../utils/AppProvider";
+import { AppContext } from "../context/AppContext";
 import {
   IconButton,
   Menu,
