+++ src/utils/wrapper/wrap-root-element.js
@@ -5,7 +5,7 @@ import React from "react";
 import { ColorModeProvider, CSSReset, ThemeProvider } from "@chakra-ui/core";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.min.css";
-import AppProvider from "../AppProvider";
+import AppProvider from "../../context/AppContext";
 import ApolloProvider from "./Apollo-Provider";
 
 export const WrapRootElement = ({ children }) => {
